Address review round 1
The compose passthrough I added for #1378 was itself an instance of the bug it
was fixing: BASELINE_LOCK_MODE_MYDUMPER was referenced inside the container but
never listed in the baseline-dump service's environment: block, so it was always
unset, the :-FTWRL default always won, and an operator's choice was silently
ignored. Both compose surfaces now take the SAME BASELINE_LOCK_MODE in the same
lowercase vocabulary as the console, it is actually passed into the container,
and an unrecognized value exits non-zero instead of falling through.

Second, and worse in kind: an invalid BINTRAIL_CONSOLE_BASELINE_LOCK_MODE made
`watch` refuse to boot. Under watch that process is also the capture plane, so a
typo in a BASELINE setting would have turned into permanently lost binlog
events -- and the compose file 280 lines away taught the uppercase spelling that
ParseLockMode rejects. The error now disables baselines only: it is carried to
the supervisor and returned from every Trigger, where the operator is looking.
Same rule as audit readability gating nothing in the capture path.

Third, the privilege refusal -- which now fires on the DEFAULT path, so it is
the first thing an upgrading least-privilege deployment sees -- named
BINTRAIL_CONSOLE_BASELINE_POINT_CONSISTENT, a variable this series deleted the
read of, and told the operator to "disable point-consistent mode", which is no
longer a thing. All four branches now name BINTRAIL_CONSOLE_BASELINE_LOCK_MODE
and say what safe-no-lock and no-lock actually trade.

Fourth, the CLI got the FTWRL default without the preflight that makes the
mydumper segfault unreachable. Granting BACKUP_ADMIN without RELOAD does not
fail cleanly -- the pinned build crashes -- and before this series the CLI
always passed NO_LOCK and could not reach it. checkPointConsistentPrivileges
moved to internal/mydumperlock and both surfaces call it, rather than the docs
claiming a parity that did not exist.

Also: lockmode.go claimed PointConsistent() is "what the footer marker keys
off" -- there is no footer marker, and that is the remaining half of #1377, now
stated as absent instead of implied. The CHANGELOG understated the blast radius
twice: FTWRL + --trx-tables makes mydumper refuse a dump outright on any
non-transactional table, which breaks fully-privileged operators too; and the
stack's own documented source user does not satisfy the new default while
BASELINE_TRIGGER defaults on. .env.example and the quickstart now say so.

Stale prose swept: docs/console.md and docs/docker.md still documented the
removed variable and NO_LOCK as the default, dump-and-baseline.md contradicted
itself eleven lines apart and omitted --lock-mode from its own flag table, and
four comment blocks plus two test names still described a pointConsistent bool
that no longer exists.

Tests: the two failure paths this round created -- daemon survives an invalid
mode, supervisor refuses the trigger -- both mutation-verified.
