+++ .env.example
@@ -9,6 +9,17 @@
 # OPTIONAL — a source MySQL to start watching immediately at boot.
 # Leave unset and add servers from the console UI instead (zero-config path).
 # The user needs REPLICATION SLAVE, REPLICATION CLIENT, and SELECT.
+# Baselines additionally need RELOAD (or FLUSH_TABLES), plus BACKUP_ADMIN on
+# MySQL/Percona 8.0+, because the snapshot is point-consistent by default.
+# Without them, set BASELINE_LOCK_MODE below instead of granting.
+#
+# BASELINE_LOCK_MODE: how the baseline dump syncs mydumper's threads onto one
+# instant. ftwrl (default) is point-consistent; safe-no-lock needs no extra
+# privilege but ABORTS rather than write a snapshot stitched from several
+# instants, so expect it to refuse on a write-active source; no-lock accepts
+# such a snapshot. A baseline is the seed state reconstruct merges deltas onto,
+# which is why the weaker modes must be asked for by name.
+#BASELINE_LOCK_MODE=ftwrl
 # `host.docker.internal` reaches a MySQL on this machine from inside Docker.
 # SOURCE_DSN=user:password@tcp(host.docker.internal:3306)/
