+++ consoleapp/daemon_beacon_test.go
@@ -29,7 +29,7 @@ func clearConsoleEnv(t *testing.T) {
 		"BINTRAIL_CONSOLE_ALLOWED_HOSTS", "BINTRAIL_CONSOLE_ALLOW_SETUP",
 		"BINTRAIL_CONSOLE_FLASHBACK_LISTEN", "BINTRAIL_CONSOLE_ARCHIVE_STAGING",
 		"BINTRAIL_CONSOLE_BASELINE_TRIGGER", "BINTRAIL_CONSOLE_BASELINE_STAGING",
-		"BINTRAIL_CONSOLE_BASELINE_POINT_CONSISTENT",
+		"BINTRAIL_CONSOLE_BASELINE_LOCK_MODE",
 		"BINTRAIL_CONSOLE_VERIFY_TRIGGER", "BINTRAIL_CONSOLE_VERIFY_INTERVAL",
 		"BINTRAIL_CONSOLE_VERIFY_TABLES", "BINTRAIL_CONSOLE_NOTIFY_WEBHOOK",
 	} {
